[NB] Enable 32-bit packing of array structures
Enable packing of array structures to avoid insertion of padding between
member variables.  This makes it possible to cast memory directly as a
structure, allowing zero overhead for deserialization.

To make this work on both 8-bit and 32-bit architectures classes need to
be [passive data structures][1] (aka., plain old data).  This implies
that classes may not have constructors or other methods.  For this
library, this means that, as of this commit, the existing constructor
methods for the `UInt8Array` class have been replaced by the following
inline functions:

    `UInt8Array_init_default()`
    `UInt8Array_init(uint32_t length, uint8_t *data)`

Also, note that, as of this commit, the `length` variable of each array
class has been changed from `uint16_t` to `uint32_t`.

[1]: https://en.wikipedia.org/wiki/Passive_data_structure